requirements.txt
requests
requests_oauthlib
py_cpuinfo
psutil
flask
flask_cors
pynput
python-telegram-bot
twilio
scapy
NetfilterQueue
boto3
geocoder
pathlib
wget
yara-python
clamd
beautifulsoup4
pyudev
